+++ tests/api/search.test.js
@@ -8,10 +8,8 @@ const squad = require('../../config').get('squadName')
 const {
   getKubeConfig,
   getSearchApiRoute,
-  getToken,
-  searchQueryBuilder,
-  sendRequest,
-} = require('../common-lib/clusterAccess')
+  getToken } = require('../common-lib/clusterAccess')
+const { searchQueryBuilder, sendRequest } = require('../common-lib/searchClient')
 
 describe('RHACM4K-913: Search - Verify search results with different queries', () => {
   // Get kubeconfig for imported clusters
