+++ tests/common-lib/searchClient.js
@@ -1,16 +1,23 @@
 // Copyright Contributors to the Open Cluster Management project
 
+/*
+ * This file has functions to interact with the Search API.
+ */
 const { fail } = require('assert')
-const { performance } = require('perf_hooks')
-const {
-    searchQueryBuilder,
-    sendRequest } = require('./clusterAccess')
-    
+const { performance } = require('perf_hooks')    
 const {
     formatResourcesFromSearch,
     formatFilters } = require('./index')
+const request = require('supertest')
 
-
+/**
+ * Query the Search API using the given filters.
+ * 
+ * @param {*} kind The kind filter
+ * @param {*} apigroup The apigroup filter
+ * @param {*} cluster The cluster fiilter.
+ * @param {*} namespace The namespace filter.
+ */
 async function getResourcesFromSearch(kind,
     apigroup,
     namespace = '--all-namespaces',
@@ -41,4 +48,50 @@ async function getResourcesFromSearch(kind,
     return formatResourcesFromSearch(resp)
   }
 
+/**
+ * Builds and returns a query object for a HTTP request. (Current supported input keys: `keywords`, `filters`, and `limit`)
+ * @param {object} {} The input keys that will be used to build the query object.
+ * @param {object} options Additional options for building the query object..
+ * @returns {object} The query object.
+ */
+ function searchQueryBuilder(
+    { keywords = [], filters = [], limit = 10000 },
+    options = {}
+  ) {
+    // Return query built from passed arguments.
+    const query = {
+      operationName: 'searchResult',
+      variables: {
+        input: [
+          {
+            keywords: keywords,
+            filters: filters,
+            limit: limit,
+          },
+        ],
+      },
+      query:
+        'query searchResult($input: [SearchInput]) {\n  searchResult: search(input: $input) {\n    items\n    __typename\n  }\n}\n',
+    }
+    return query
+}
+
+/**
+ * Send a HTTP request to the API server and return the results. Expects the response to have a 200 status code.
+ * @param {*} query The query to send.
+ * @param {*} token The validation token to use for the request.
+ * @param {object} options Additional options for sending the request.
+ * @returns
+ */
+ function sendRequest(query, token, options = {}) {
+    return request(searchApiRoute)
+      .post('/searchapi/graphql')
+      .send(query)
+      .set({ Authorization: `Bearer ${token}` })
+      .expect(200)
+  }
+  
 exports.getResourcesFromSearch = getResourcesFromSearch
+exports.searchQueryBuilder = searchQueryBuilder
+exports.sendRequest = sendRequest
+
